KAN/requirements.txt
pykan
pytorch
numpy
matplotlib
